Put more files in place for organized graphics development
