+++ src/components/game/graphics/objects/cursor/DrawCursor.js
@@ -1,6 +1,6 @@
-import { outlineRect } from "../utils/Shapes";
+import { outlineRect } from "../../utils/Shapes";
 
-export const renderCursor = (canvas, board, controller, windowSize, xCellSize, yCellSize) => {
+export const drawCursor = (canvas, board, controller, windowSize, xCellSize, yCellSize) => {
     var [x, y] = controller.gridCursor(
         windowSize,
         board.length
